requirements.txt
amqp==5.1.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
beautifulsoup4==4.12.3
billiard==4.1.0
bs4==0.0.2
celery==5.3.4
celery_once==3.0.1
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorlog==6.7.0
cron_descriptor==1.4.0
cssselect==1.2.0
Django==4.2.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-db-geventpool==4.0.5
django-filter==23.2
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-ratelimit==4.1.0
django-redis==5.3.0
django-simple-captcha==0.6.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
fake-useragent==1.5.1
gunicorn==22.0.0
h11==0.14.0
idna==3.7
importlib_metadata==7.1.0
inflection==0.5.1
kombu==5.3.2
lxml==5.2.2
lxml_html_clean==0.1.1
msgpack==1.0.8
packaging==23.1
parse==1.20.1
pillow==10.3.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.5
pycryptodomex==3.20.0
pyee==11.1.0
PyMySQL==1.1.0
pyppeteer==2.0.0
pyquery==2.0.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzipper==0.3.6
redis==4.6.0
reportlab==4.2.2
requests==2.32.3
requests-html==0.10.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.4.4
tqdm==4.66.4
typing_extensions==4.12.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.23.2
vine==5.0.0
w3lib==2.1.2
wcwidth==0.2.6
websockets==10.4
zipp==3.19.1
